+++ core/synthesis/structure_synthesis/__init__.py
@@ -44,7 +44,6 @@ from core.QtModules import (
     QHeaderView,
 )
 from core.libs import (
-    VJoint,
     Graph,
     link_assortment,
     contracted_link_assortment,
@@ -175,32 +174,30 @@ class StructureSynthesis(QWidget, Ui_Form):
 
     @Slot(name='on_from_mechanism_button_clicked')
     def __from_mechanism(self):
-        """Reload button: Auto-combine the mechanism from the workbook."""
+        """From a generalized mechanism of main canvas."""
         if self.vpoints and self.vlinks:
             graph, _, _, _, _, _ = self.get_graph()
-            self.edges_text.setText(str(graph.edges))
         else:
             graph = Graph([])
+        if graph.edges:
+            self.edges_text.setText(str(list(graph.edges)))
+        else:
             self.edges_text.setText("")
         keep_dof_checked = self.keep_dof.isChecked()
         self.keep_dof.setChecked(False)
-        self.nl_input.setValue(
-            sum(len(vlink.points) > 1 for vlink in self.vlinks) +
-            sum(
-                len(vpoint.links) - 2 for vpoint in self.vpoints
-                if vpoint.type == VJoint.RP and len(vpoint.links) > 1
-            )
-        )
-        self.nj_input.setValue(sum(
-            (len(vpoint.links) - 1 + int(vpoint.type == VJoint.RP))
-            for vpoint in self.vpoints if len(vpoint.links) > 1
-        ))
+        self.nl_input.setValue(len(graph.nodes))
+        self.nj_input.setValue(len(graph.edges))
         self.keep_dof.setChecked(keep_dof_checked)
 
-        # Auto synthesis
-        if not graph.edges:
-            return
-        self.__number_synthesis()
+        # Show attributes
+        QMessageBox.information(
+            self,
+            "Generalization",
+            f"Link assortment:\n{link_assortment(graph)}\n"
+            f"Contracted link assortment:\n{contracted_link_assortment(graph)}"
+            if graph.edges else
+            "Is a empty graph."
+        )
 
     def __adjust_structure_data(self):
         """Update NJ and NL values.