+++ docs/changelog.md
@@ -6,6 +6,7 @@
 + Change the behavior of adding points on main canvas.
 + Add duplicate function for graph collections page.
 + Fix the error of switching structure collections.
++ Showing attributes instead doing number synthesis after generalization.
 
 **Development**: